Rename BluetoothHelper to BluetoothController, pair with nfc, multiple bt module support, BluetoothControllerDelegate
