--- smart-library-app/BluetoothHelper.swift
@@ -1,160 +1,0 @@
-//
-//  BluetoothHelper.swift
-//  smart-library-app
-//
-//  Created by Riley Keane on 10/14/22.
-//
-
-import Foundation
-import CoreBluetooth
-
-private struct CBUUIDs{
-    
-    static let kBLEService_UUID = "0000ffe0-0000-1000-8000-00805f9b34fb"
-    static let kBLE_Characteristic_uuid_Tx = "0000ffe1-0000-1000-8000-00805f9b34fb"
-    static let kBLE_Characteristic_uuid_Rx = "0000ffe1-0000-1000-8000-00805f9b34fb"
-
-    static let BLEService_UUID = CBUUID(string: kBLEService_UUID)
-    static let BLE_Characteristic_uuid_Tx = CBUUID(string: kBLE_Characteristic_uuid_Tx)//(Property = Write without response)
-    static let BLE_Characteristic_uuid_Rx = CBUUID(string: kBLE_Characteristic_uuid_Rx)// (Property = Read/Notify)
-}
-
-class BluetoothHelper: NSObject, CBPeripheralDelegate, CBCentralManagerDelegate {
-    static let shared = BluetoothHelper()
-    
-    private var centralManager: CBCentralManager!
-    private var peripheral: CBPeripheral!
-    
-    private var txCharacteristic: CBCharacteristic!
-    private var rxCharacteristic: CBCharacteristic!
-    
-    private var deviceToConnectTo: UUID?
-    
-    override private init() {
-        super.init()
-        centralManager = CBCentralManager(delegate: self, queue: nil)
-    }
-    
-    func centralManagerDidUpdateState(_ central: CBCentralManager) {
-        switch central.state {
-                  case .poweredOff:
-                      print("Is Powered Off.")
-                  case .poweredOn:
-                      print("Is Powered On.")
-//                        startScanning(btDeviceUuid: UUID(uuidString: "8A31C81E-FBA7-DB87-45BE-D53ADA6CCFC7"))
-                  case .unsupported:
-                      print("Is Unsupported.")
-                  case .unauthorized:
-                  print("Is Unauthorized.")
-                  case .unknown:
-                      print("Unknown")
-                  case .resetting:
-                      print("Resetting")
-                  @unknown default:
-                    print("Error")
-                }
-    }
-    
-    func startScanning(btDeviceUuid: UUID?) -> Void {
-      // Start Scanning
-        centralManager?.scanForPeripherals(withServices: [CBUUIDs.BLEService_UUID])
-        self.deviceToConnectTo = btDeviceUuid
-        
-    }
-    
-    func centralManager(_ central: CBCentralManager, didDiscover peripheral: CBPeripheral,advertisementData: [String : Any], rssi RSSI: NSNumber) {
-
-        self.peripheral = peripheral
-        self.peripheral.delegate = self
-
-        print("Peripheral Discovered: \(peripheral.identifier)")
-        print("Peripheral name: \(peripheral.name)")
-//        print ("Advertisement Data : \(advertisementData)")
-        
-        if (peripheral.identifier == self.deviceToConnectTo) {
-            centralManager?.connect(self.peripheral, options: nil)
-        }
-        
-    }
-    
-    func centralManager(_ central: CBCentralManager, didConnect peripheral: CBPeripheral) {
-        self.peripheral.discoverServices([CBUUIDs.BLEService_UUID])
-    }
-    
-    func peripheral(_ peripheral: CBPeripheral, didDiscoverServices error: Error?) {
-//            print("*******************************************************")
-
-            if ((error) != nil) {
-                print("Error discovering services: \(error!.localizedDescription)")
-                return
-            }
-            guard let services = peripheral.services else {
-                return
-            }
-            //We need to discover the all characteristic
-            for service in services {
-                peripheral.discoverCharacteristics(nil, for: service)
-            }
-//        print("Discovered Services: \(services)")
-    }
-    
-    func peripheral(_ peripheral: CBPeripheral, didDiscoverCharacteristicsFor service: CBService, error: Error?) {
-           
-               guard let characteristics = service.characteristics else {
-              return
-          }
-
-          print("Found \(characteristics.count) characteristics.")
-
-          for characteristic in characteristics {
-
-            if characteristic.uuid.isEqual(CBUUIDs.BLE_Characteristic_uuid_Rx)  {
-
-              rxCharacteristic = characteristic
-
-              peripheral.setNotifyValue(true, for: rxCharacteristic!)
-              peripheral.readValue(for: characteristic)
-
-              print("RX Characteristic: \(rxCharacteristic.uuid)")
-            }
-
-            if characteristic.uuid.isEqual(CBUUIDs.BLE_Characteristic_uuid_Tx){
-              
-              txCharacteristic = characteristic
-              
-              print("TX Characteristic: \(txCharacteristic.uuid)")
-            }
-          }
-    }
-    
-    func peripheral(_ peripheral: CBPeripheral, didUpdateValueFor characteristic: CBCharacteristic, error: Error?) {
-        
-        print(characteristic)
-        
-        var characteristicASCIIValue = NSString()
-
-        guard characteristic == rxCharacteristic,
-                let characteristicValue = characteristic.value,
-                let ASCIIstring = NSString(data: characteristicValue, encoding: String.Encoding.utf8.rawValue) else {
-            return
-        }
-        characteristicASCIIValue = ASCIIstring
-        
-//        print("Value Recieved: \((characteristicASCIIValue as String))")
-    }
-    
-    func writeOutgoingValue(data: String){
-          
-        let valueString = (data as NSString).data(using: String.Encoding.utf8.rawValue)
-        
-        if let peripheral = self.peripheral {
-              
-          if let txCharacteristic = txCharacteristic {
-              print("sending message")
-              self.peripheral.writeValue(valueString!, for: txCharacteristic, type: CBCharacteristicWriteType.withResponse)
-              }
-          }
-      }
-}
-
-
